Move Encoder and Gear, update imports
