+++ src/main/java/com/stuypulse/robot/commands/DrivetrainCommand.java
@@ -5,7 +5,6 @@
 package com.stuypulse.robot.commands;
 
 import com.stuypulse.robot.subsystems.Drivetrain;
-import com.stuypulse.robot.util.gear.Gear;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
@@ -31,7 +30,7 @@ public abstract class DrivetrainCommand extends CommandBase {
 
     public abstract double getAngle();
 
-    public abstract Gear getGear();
+    public abstract Drivetrain.Gear getGear();
 
     public abstract boolean useCurvatureDrive();
